backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pyyaml>=6.0
requests>=2.32.0
pydantic>=2.0
supabase>=2.0.0
